simplify PageController
